cassandra_test/requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
cached-property==1.3.1
cassandra-pylib==0.0.0
certifi==2018.1.18
chardet==3.0.4
Cheetah==2.4.4
cryptography==2.1.4
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
gyp==0.1
html5lib==0.999999999
idna==2.6
ipaddr==2.2.0
ipaddress==1.0.17
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
libvirt-python==4.0.0
lxml==4.2.1
M2Crypto==0.27.0
mock==2.0.0
numpy==1.13.3
pbr==3.1.1
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
PySocks==1.7.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
scour==0.36
SecretStorage==2.3.1
six==1.11.0
texttable==0.9.1
typing==3.6.2
urllib3==1.22
webencodings==0.5
websocket-client==0.44.0
